# TraceQL performance and parquetquery cleanup